backend/user_service/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==5.0.2
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jwcrypto==1.5.3
kombu==5.3.5
oauthlib==3.2.2
phonenumbers==8.13.30
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
sendgrid==6.8.0
gunicorn==21.2.0
pyotp==2.9.0